+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Updated CHANGELOG.md
   - Renamed NetworkingDSc to NetworkingDsc in CHANGELOG.md - fixes [Issue #513](https://github.com/dsccommunity/NetworkingDsc/issues/513).
+- NetAdapterName
+  - Added new parameter HyperVNetworkAdapterName to filter if device naming is configured on a Hyper-V network adapter.
 
 ## [9.0.0] - 2022-05-30